戻り値の型をintで定義しました。

+++ src/Review01.java
@@ -5,14 +5,14 @@ public class Review01 {
 		int price = 1500;
 		int taxRate = 10;
 
-		int tax = (int) tax(price, taxRate);
+		int tax = tax(price, taxRate);
 		System.out.println ( price + "円の商品の税込価格は" + (price + tax) + "円（消費税は" + tax + "円）です。");
 
 	}
 
-	public static double tax(double price,double taxRate) {
+	public static int tax(int price,int taxRate) {
 
-		double tax = price * (taxRate/100);
-		return tax;
-	}
+		int tax = price * taxRate/100;
+		return  tax;
+}
 }
